Revert "Revert "Changed MouseMoveEvent from doubles to ints""
This reverts commit 307d7a7315dbc17a5f85bc5b7ca9ff507f11eb58.

+++ GameProject/Engine/Events/Events.h
@@ -30,11 +30,11 @@ struct WindowResizeEvent : public Event
 
 struct MouseMoveEvent : public Event
 {
-	MouseMoveEvent(double moveX, double moveY, double travelX, double travelY) : moveX{ moveX }, moveY{ moveY }, travelX{ travelX }, travelY{ travelY } {};
-	double moveX;
-	double moveY;
-	double travelX;
-	double travelY;
+	MouseMoveEvent(int moveX, int moveY, int travelX, int travelY) : moveX{ moveX }, moveY{ moveY }, travelX{ travelX }, travelY{ travelY } {};
+	int moveX;
+	int moveY;
+	int travelX;
+	int travelY;
 };
 
 struct MouseClickEvent : public Event
